run ingestor in Lambda
