+++ ingestor/src/main/resources/application.conf
@@ -18,4 +18,4 @@ data-target.cluster = {
 }
 
 data-target.index.name=authoritative-concepts
-data-target.bulk.max-records=50000
+data-target.bulk.max-records=10000
